Updated to allow event listeners to return data to the caller of the emit() function, e.g. `const results = emitter.emit("myEventName");`
The results are an array of all return values from any listener functions that were called by the event "myEventName". This is useful if you want event listeners to act as data providers for particular events.

Listeners that return `ListenerReturnFlag.cancel` will cancel the event call to any further listeners for that one event emit.